Fixed mapping employees to dashboard employees list and added functionality to display a newly created employee onto the list

+++ Capstone Project/server/src/main/java/com/server/Capstone/Project/models/UpdateUserDTO.java
@@ -7,7 +7,7 @@ public class UpdateUserDTO {
     private String dateOfBirth;
     private String address;
     private String phoneNumber;
-    private int proRata;
+    private double proRata;
     private int salary;
 
     public UpdateUserDTO(){
@@ -62,11 +62,11 @@ public class UpdateUserDTO {
         this.phoneNumber = phoneNumber;
     }
 
-    public int getProRata() {
+    public double getProRata() {
         return proRata;
     }
 
-    public void setProRata(int proRata) {
+    public void setProRata(double proRata) {
         this.proRata = proRata;
     }
